+++ sys/Kconfig
@@ -17,6 +17,7 @@ rsource "entropy_source/Kconfig"
 rsource "event/Kconfig"
 rsource "fmt/Kconfig"
 rsource "isrpipe/Kconfig"
+rsource "malloc_thread_safe/Kconfig"
 rsource "net/Kconfig"
 rsource "Kconfig.newlib"
 rsource "Kconfig.stdio"
