+++ sys/malloc_thread_safe/Makefile
@@ -1,0 +1,1 @@
+include $(RIOTBASE)/Makefile.base
